requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Task Queue & Cache
celery>=5.3.4
redis>=5.0.0

# ML/NLP
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
langchain>=0.0.300
google-generativeai>=0.3.0

# Document Parsing (Production)
unstructured>=0.11.0
pypdf>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.0

# PII Redaction (GDPR)
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.42.0

# Utilities
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Deployment
gunicorn>=21.0.0